Cerrar el resto de P1: industrias, casas, órdenes y horarios.
Transformación de insumos temperate, pax/correo por HouseSpec,
flags de orden (no_load/full_load_any/non_stop/transfer), horarios
reales con autofill y lateness; cliente y clippy al día. P1 22/22.

+++ crates/openttdrs-core/src/cargo_packet/types.rs
@@ -309,8 +309,7 @@ impl VehicleCargoList {
                 let mut taken = p.clone();
                 taken.count = take;
                 if p.feeder_share > 0 {
-                    let share = (i64::from(p.feeder_share) * i64::from(take))
-                        / i64::from(p.count);
+                    let share = (p.feeder_share * i64::from(take)) / i64::from(p.count);
                     taken.feeder_share = share;
                     p.feeder_share = p.feeder_share.saturating_sub(share);
                 }
